theme update
-> worked on the custom menu images for the recipe tabs
If you want to change the image of a recipe in the mobile menu, you can do so in the customization options under the header tab. However, you must make sure that the recipe name remains the same as it is in the recipe menu. The screenshot below shows you how to do this.
https://prnt.sc/m8rqPevKgr-f
https://prnt.sc/vJuzfK4NA-26

-> worked on the loading issue of recipe page